@@ -26,14 +26,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.0.1",
-    "@angular/common": "^19.0.1",
-    "@angular/compiler": "^19.0.1",
-    "@angular/core": "^19.0.1",
-    "@angular/forms": "^19.0.1",
-    "@angular/platform-browser": "^19.0.1",
-    "@angular/platform-browser-dynamic": "^19.0.1",
-    "@angular/router": "^19.0.1",
+    "@angular-devkit/build-angular": "^20.0.0",
+    "@angular/animations": "^20.0.0",
+    "@angular/common": "^20.0.0",
+    "@angular/compiler": "^20.0.0",
+    "@angular/core": "^20.0.0",
+    "@angular/forms": "^20.0.0",
+    "@angular/platform-browser": "^20.0.0",
+    "@angular/platform-browser-dynamic": "^20.0.0",
+    "@angular/router": "^20.0.0",
     "core-js": "^3.26.1",
     "rxjs": "~7.8.1",
     "svg-to-ts": "^12.0.0",
@@ -41,17 +42,17 @@
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.0.2",
-    "@angular-devkit/core": "^19.0.2",
-    "@angular-devkit/schematics": "^19.0.2",
-    "@angular-eslint/builder": "^19.0.0",
-    "@angular-eslint/eslint-plugin": "^19.0.0",
-    "@angular-eslint/eslint-plugin-template": "^19.0.0",
-    "@angular-eslint/schematics": "^19.0.0",
-    "@angular-eslint/template-parser": "^19.0.0",
-    "@angular/cli": "^19.0.2",
-    "@angular/compiler-cli": "^19.0.1",
-    "@angular/language-service": "^19.0.1",
+    "@angular-devkit/core": "^20.0.0",
+    "@angular-devkit/schematics": "^20.0.0",
+    "@angular-eslint/builder": "^19.6.0",
+    "@angular-eslint/eslint-plugin": "^19.6.0",
+    "@angular-eslint/eslint-plugin-template": "^19.6.0",
+    "@angular-eslint/schematics": "^19.6.0",
+    "@angular-eslint/template-parser": "^19.6.0",
+    "@angular/build": "^20.0.0",
+    "@angular/cli": "^20.0.0",
+    "@angular/compiler-cli": "^20.0.0",
+    "@angular/language-service": "^20.0.0",
     "@types/jasmine": "~5.1.5",
     "@types/jasminewd2": "~2.0.13",
     "@types/node": "^22.10.1",
@@ -70,10 +71,10 @@
     "karma-coverage": "^2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "^2.0.0",
-    "ng-packagr": "^19.0.1",
+    "ng-packagr": "^20.0.0",
     "rxjs-marbles": "^7.0.1",
     "ts-node": "~10.9.1",
-    "typescript": "~5.6.3",
+    "typescript": "~5.8.3",
     "webpack-bundle-analyzer": "^4.7.0"
   }
 }
