@@ -1,9 +1,9 @@
 {
   "name": "@triangular/noto-emoji-icons",
-  "version": "11.0.0",
+  "version": "20.0.0",
   "peerDependencies": {
-    "@angular/common": "^19.0.0",
-    "@angular/core": "^19.0.0"
+    "@angular/common": "^20.0.0",
+    "@angular/core": "^20.0.0"
   },
   "homepage": "https://noto-emoji.akehir.com",
   "license" : "MIT",
