@@ -2,7 +2,6 @@ import { Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { UntypedFormControl } from '@angular/forms';
 import { filter, map, throttleTime } from 'rxjs/operators';
 
-// eslint-disable-next-line @angular-eslint/prefer-standalone
 @Component({
   selector: 'app-search',
   template: `
@@ -10,22 +9,22 @@ import { filter, map, throttleTime } from 'rxjs/operators';
     <div class="search__query">
       <label class="search__label" for="search">Search</label>
       <input class="search__field"
-             type="text"
-             id="search"
-             [formControl]="term"
-             placeholder="Search"
-             autocomplete="off">
+        type="text"
+        id="search"
+        [formControl]="term"
+        placeholder="Search"
+        autocomplete="off">
     </div>
     <div class="search__answer">
-      <ng-container *ngFor="let result of results">
+      @for (result of results; track result) {
         <a [routerLink]="result.slice(0,1)" [fragment]="result" (click)="clear()">{{result}}</a>
-      </ng-container>
+      }
     </div>
   </div>
-`,
+  `,
   styleUrls: [],
   encapsulation: ViewEncapsulation.None,
-  standalone: false
+  standalone: false // eslint-disable-line @angular-eslint/prefer-standalone
 })
 export class SearchComponent implements OnInit {
   term = new UntypedFormControl('');
