@@ -120,13 +120,12 @@ import {
   notoEmojiU1F390,
 } from '@triangular/noto-emoji-icons/icons';
 
-// eslint-disable-next-line @angular-eslint/prefer-standalone
 @Component({
   selector: 'app-icons-b',
   templateUrl: './icons-b.component.html',
   styleUrls: [],
   encapsulation: ViewEncapsulation.None,
-  standalone: false
+  standalone: false // eslint-disable-line @angular-eslint/prefer-standalone
 })
 export class IconsBComponent {
   icons = [
