@@ -120,13 +120,12 @@ import {
   notoEmojiU1F4821F3Fc200D2642,
 } from '@triangular/noto-emoji-icons/icons';
 
-// eslint-disable-next-line @angular-eslint/prefer-standalone
 @Component({
   selector: 'app-icons-j',
   templateUrl: './icons-j.component.html',
   styleUrls: [],
   encapsulation: ViewEncapsulation.None,
-  standalone: false
+  standalone: false // eslint-disable-line @angular-eslint/prefer-standalone
 })
 export class IconsJComponent {
   icons = [
