@@ -120,13 +120,12 @@ import {
   notoEmojiU1F9391F3Fe200D2640,
 } from '@triangular/noto-emoji-icons/icons';
 
-// eslint-disable-next-line @angular-eslint/prefer-standalone
 @Component({
   selector: 'app-icons-r',
   templateUrl: './icons-r.component.html',
   styleUrls: [],
   encapsulation: ViewEncapsulation.None,
-  standalone: false
+  standalone: false // eslint-disable-line @angular-eslint/prefer-standalone
 })
 export class IconsRComponent {
   icons = [
