@@ -120,13 +120,12 @@ import {
   notoEmojiU1F9D11F3Fc200D1F9B2,
 } from '@triangular/noto-emoji-icons/icons';
 
-// eslint-disable-next-line @angular-eslint/prefer-standalone
 @Component({
   selector: 'app-icons-u',
   templateUrl: './icons-u.component.html',
   styleUrls: [],
   encapsulation: ViewEncapsulation.None,
-  standalone: false
+  standalone: false // eslint-disable-line @angular-eslint/prefer-standalone
 })
 export class IconsUComponent {
   icons = [
